# Conflicts:
#	analysis_file/analysis_file.py

+++ messages/messages.py
@@ -1,11 +1,11 @@
 import argparse
+import os
 import time
 
 import pytz
 from core_data_modules.cleaners import somali
 from core_data_modules.traced_data import Metadata
-from core_data_modules.traced_data.io import TracedDataJsonIO, TracedDataCodaIO, \
-    TracedDataTheInterfaceIO
+from core_data_modules.traced_data.io import TracedDataJsonIO, TracedDataCodaIO
 from core_data_modules.util import IOUtils
 from dateutil.parser import isoparse
 
@@ -15,24 +15,24 @@ if __name__ == "__main__":
     parser.add_argument("user", help="User launching this program")
     parser.add_argument("json_input_path", metavar="json-input-path",
                         help="Path to the input JSON file, containing a list of serialized TracedData objects")
+    parser.add_argument("prev_coda_path", metavar="prev-coda-path",
+                        help="Path to a Coda file containing previously coded messages")
     parser.add_argument("flow_name", metavar="flow-name",
                         help="Name of activation flow from which this data was derived")
     parser.add_argument("variable_name", metavar="variable-name",
                         help="Name of message variable in flow")
     parser.add_argument("json_output_path", metavar="json-output-path",
                         help="Path to a JSON file to write processed messages to")
-    parser.add_argument("interface_output_dir", metavar="interface-output-dir",
-                        help="Path to a directory to write The Interface files to")
     parser.add_argument("coda_output_path", metavar="coda-output-path",
                         help="Path to a Coda file to write processed messages to")
 
     args = parser.parse_args()
     user = args.user
     json_input_path = args.json_input_path
+    prev_coda_path = args.prev_coda_path
     variable_name = args.variable_name
     flow_name = args.flow_name
     json_output_path = args.json_output_path
-    interface_output_dir = args.interface_output_dir
     coda_output_path = args.coda_output_path
 
     # Load data from JSON file
@@ -47,7 +47,7 @@ if __name__ == "__main__":
     show_messages = [td for td in show_messages if show_message_key in td]
 
     # Filter out messages containing only noise
-    print("Messages classed as noise:")
+    print("Messages classified as noise:")
     not_noise = []
     for td in show_messages:
         if somali.DemographicCleaner.is_noise(td[show_message_key]):
@@ -73,13 +73,15 @@ if __name__ == "__main__":
     with open(json_output_path, "w") as f:
         TracedDataJsonIO.export_traced_data_iterable_to_json(show_messages, f, pretty_print=True)
 
-    # Output to The Interface
-    IOUtils.ensure_dirs_exist(interface_output_dir)
-    TracedDataTheInterfaceIO.export_traced_data_iterable_to_the_interface(
-        show_messages, interface_output_dir, "avf_phone_id", show_message_key, eat_key)
-
     # Output messages to Coda
     IOUtils.ensure_dirs_exist_for_file(coda_output_path)
-    with open(coda_output_path, "w") as f:
-        TracedDataCodaIO.export_traced_data_iterable_to_coda(
-            show_messages, "{} (Text) - {}".format(variable_name, flow_name), f)
+    if os.path.exists(prev_coda_path):
+        # TODO: Modifying this line once the coding frame has been developed to include lots of Nones feels a bit
+        # TODO: cumbersome. We could instead modify export_traced_data_iterable_to_coda to support a prev_f argument.
+        scheme_keys = {"Relevance": None, "Code 1": None, "Code 2": None, "Code 3": None, "Code 4": None}
+        with open(coda_output_path, "w") as f, open(prev_coda_path, "r") as prev_f:
+            TracedDataCodaIO.export_traced_data_iterable_to_coda_with_scheme(
+                show_messages, show_message_key, scheme_keys, f, prev_f=prev_f)
+    else:
+        with open(coda_output_path, "w") as f:
+            TracedDataCodaIO.export_traced_data_iterable_to_coda(show_messages, show_message_key, f)
